# Conflicts:
#	app/src/main/java/com/ssttkkl/fgoplanningtool/resources/ResourcesProvider.kt
#	app/src/main/java/com/ssttkkl/fgoplanningtool/ui/costitemlist/CostItemListEntity.kt

+++ app/src/main/java/com/ssttkkl/fgoplanningtool/data/item/CostItemsExtension.kt
@@ -5,7 +5,7 @@ import com.ssttkkl.fgoplanningtool.data.plan.Plan
 private const val HOLY_GRAIL = "holy_grail"
 private const val QP = "qp"
 
-private fun MutableMap<String, Int>.putItems(items: Collection<Item>) {
+private fun MutableMap<String, Long>.putItems(items: Collection<Item>) {
     items.forEach { item ->
         val oldValue = get(item.codename) ?: 0
         this[item.codename] = oldValue + item.count
@@ -14,7 +14,7 @@ private fun MutableMap<String, Int>.putItems(items: Collection<Item>) {
 
 val Collection<Plan>.costItems: Collection<Item>
     get() {
-        val map = HashMap<String, Int>()
+        val map = HashMap<String, Long>()
         forEach { plan -> map.putItems(plan.costItems) }
         return map.map { Item(it.key, it.value) }
     }
@@ -23,9 +23,9 @@ val Plan.costItems: Collection<Item>
     get() {
         if (servant == null)
             return listOf()
-        val map = HashMap<String, Int>()
-        var qp = 0
-        var holyGrail = 0
+        val map = HashMap<String, Long>()
+        var qp = 0L
+        var holyGrail = 0L
 
         for (cur in nowStage until minOf(planStage, 4)) {
             map.putItems(servant!!.ascensionItems[cur])
