+++ app/src/main/java/com/ssttkkl/fgoplanningtool/resources/ResourcesProvider.kt
@@ -46,15 +46,15 @@ class ResourcesProvider(context: Context) {
         itemRank = list.mapIndexed { idx, it -> Pair(it.codename, idx) }.toMap(HashMap())
     }
 
-    val ascensionQPInfo: List<List<Int>>
-    val skillQPInfo: List<List<Int>>
-    val palingenesisQPInfo: List<List<Int>>
+    val ascensionQPInfo: List<List<Long>>
+    val skillQPInfo: List<List<Long>>
+    val palingenesisQPInfo: List<List<Long>>
 
     init {
         val map = buildQPInfo()
-        ascensionQPInfo = map[ASCENSION] ?: List(6) { List(4) { 0 } }
-        skillQPInfo = map[SKILL] ?: List(6) { List(9) { 0 } }
-        palingenesisQPInfo = map[PALINGENESIS] ?: List(6) { List(10) { 0 } }
+        ascensionQPInfo = map[ASCENSION] ?: List(6) { List(4) { 0L } }
+        skillQPInfo = map[SKILL] ?: List(6) { List(9) { 0L } }
+        palingenesisQPInfo = map[PALINGENESIS] ?: List(6) { List(10) { 0L } }
     }
 
     private fun buildResPackInfo(): ResPackInfo {
@@ -90,10 +90,10 @@ class ResourcesProvider(context: Context) {
         }
     }
 
-    private fun buildQPInfo(): Map<String, List<List<Int>>> {
+    private fun buildQPInfo(): Map<String, List<List<Long>>> {
         try {
             File(resourcesDir, FILENAME_QP_INFO).bufferedReader().use { reader ->
-                return gson.fromJson(reader, object : TypeToken<Map<String, List<List<Int>>>>() {}.type)
+                return gson.fromJson(reader, object : TypeToken<Map<String, List<List<Long>>>>() {}.type)
             }
         } catch (exc: Exception) {
             Log.d("ResProvider", "Failed to build QPInfo map.")
