+++ app/src/main/java/com/ssttkkl/fgoplanningtool/ui/servantdetaillist/ServantDetailDialogFragment.kt
@@ -62,7 +62,7 @@ class ServantDetailDialogFragment : DialogFragment() {
     private fun generateEntities(getStart: (Int) -> String,
                                  getTo: (Int) -> String,
                                  getItems: (Servant?) -> List<Collection<Item>>,
-                                 getQP: (Servant?) -> List<Int>): List<ServantDetailLevelListEntity> {
+                                 getQP: (Servant?) -> List<Long>): List<ServantDetailLevelListEntity> {
         val qp = getQP(servant)
         val items = getItems(servant)
         return items.mapIndexed { idx, it -> it + Item("qp", qp[idx]) }
